+++ api/pool/session.go
