+++ api/router.go
@@ -4,8 +4,8 @@ import (
 	"net/http"
 
 	"github.com/gorilla/mux"
-	"github.com/minio/minio/auth"
-	"github.com/minio/minio/neofs/metrics"
+	"github.com/nspcc-dev/neofs-s3-gate/api/metrics"
+	"github.com/nspcc-dev/neofs-s3-gate/auth"
 	"go.uber.org/zap"
 )
 
@@ -79,7 +79,6 @@ type (
 )
 
 const (
-
 	// SlashSeparator - slash separator.
 	SlashSeparator = "/"
