+++ cmd/gate/app.go
@@ -7,11 +7,11 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/minio/minio/auth"
-	"github.com/minio/minio/neofs/api"
-	"github.com/minio/minio/neofs/api/handler"
-	"github.com/minio/minio/neofs/layer"
-	"github.com/minio/minio/neofs/pool"
+	"github.com/nspcc-dev/neofs-s3-gate/api"
+	"github.com/nspcc-dev/neofs-s3-gate/api/handler"
+	"github.com/nspcc-dev/neofs-s3-gate/api/layer"
+	"github.com/nspcc-dev/neofs-s3-gate/api/pool"
+	"github.com/nspcc-dev/neofs-s3-gate/auth"
 	"github.com/spf13/viper"
 	"go.uber.org/zap"
 	"google.golang.org/grpc/keepalive"
