+++ go.mod
@@ -1,12 +1,11 @@
-module github.com/minio/minio
+module github.com/nspcc-dev/neofs-s3-gate
 
-go 1.13
+go 1.14
 
 require (
 	github.com/google/uuid v1.1.1
 	github.com/gorilla/mux v1.7.4
 	github.com/klauspost/compress v1.10.10
-	github.com/minio/minio-go/v7 v7.0.3
 	github.com/nspcc-dev/neofs-api-go v1.3.0
 	github.com/nspcc-dev/neofs-crypto v0.3.0
 	github.com/pkg/errors v0.9.1
@@ -15,5 +14,11 @@ require (
 	github.com/spf13/viper v1.7.0
 	go.uber.org/atomic v1.6.0
 	go.uber.org/zap v1.15.0
+	golang.org/x/crypto v0.0.0-20200709230013-948cd5f35899 // indirect
+	golang.org/x/net v0.0.0-20200707034311-ab3426394381 // indirect
+	golang.org/x/sys v0.0.0-20200625212154-ddb9806d33ae // indirect
+	golang.org/x/text v0.3.3 // indirect
 	google.golang.org/grpc v1.29.1
+	gopkg.in/ini.v1 v1.57.0 // indirect
+	gopkg.in/yaml.v2 v2.2.8 // indirect
 )
